requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
openai==1.5.0
requests==2.31.0